refactor: use FHIR datatype tree
